Word of Wonders
Can’s CTIS 255 - Frontend Web Technologies Term Project This game is a simple version of “Words of Wonders” mobile game, which includes HTML, CSS, Javascript (jQuery). The aim is, with the given letters, finding words that exist in the puzzle. There are three main parts in the game. At the bottom of the page, there is a widget to enter letters of the words. By clicking the mouse letter button, a letter is selected, and appended to the current word. The selected letter is highlighted and cannot be selected again. In the middle of the widget, an shuffle icon is shown. If you click on it, it shuffles the leGers. After starting a word, you cannot shuffle the leGers. If you enter a word in the solution, it shows in the puzzle part.